# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend Go + Gin untuk aplikasi video conference berbasis LiveKit.\nAuth pakai dua jenis JWT: app JWT (untuk endpoint backend ini) dan LiveKit JWT (untuk connect ke LiveKit server, di-issue oleh /api/token).",
        "title": "Videoconf Backend API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/rooms": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User yang login otomatis jadi owner. Slug optional — kalau kosong, di-generate random shortuuid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Bikin room baru",
                "parameters": [
                    {
                        "description": "room data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.createRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Room"
                        }
                    },
                    "400": {
                        "description": "name kosong atau slug pattern invalid",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "409": {
                        "description": "slug udah dipake",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return semua room dimana current user adalah owner, urut paling baru duluan.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "List room milik saya",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.roomsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{idOrSlug}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Detail satu room. Kalau private, cuma owner yang bisa lihat. idOrSlug bisa numeric room id atau slug string.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Detail room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id (number) atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Room"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "private dan bukan owner",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Cascading delete: messages dan recordings ikut kehapus (FK ON DELETE CASCADE).",
                "tags": [
                    "rooms"
                ],
                "summary": "Hapus room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "bukan owner",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.createRoomRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                }
            }
        },
...
        "internal_handlers.roomsListResponse": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "created_at": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "is_locked": {
                                "type": "boolean"
                            },
                            "is_public": {
                                "type": "boolean"
                            },
                            "name": {
                                "type": "string"
                            },
                            "owner_id": {
                                "type": "integer"
                            },
                            "slug": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "videoconf-backend_internal_models.Room": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_locked": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" lalu spasi, lalu app JWT yang didapat dari /auth/login atau /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}